feat(quantity): số lượng thập phân cho hàng cân ký phía máy chủ (GL-07, POS-09)
- ADR-0015: cột số lượng và tồn kho numeric(14,3), cờ bán số lẻ trên mặt hàng và đơn vị quy đổi,
  một quy tắc tiền dòng round_half_up(đơn giá × số lượng) dùng chung máy chủ và PGlite.
- Migration 0062 đổi kiểu bằng USING col::numeric(14, 3), giữ dữ liệu.
- utils/quantity.ts: cộng trừ trên nghìn đơn vị, tiền × số lượng bằng BigInt; drizzle đọc numeric
  ở một chỗ (kiểu cột quantity()).
- Bán, trả, hủy đơn, phiếu nhập, trả hàng nhập, kiểm kê, điều chỉnh tồn, báo cáo đều nhận số lẻ và
  kiểm cờ; hàng không bật cờ gửi số lẻ trả 400.
- Nhập KiotViet tự bật cờ cho hàng đơn vị đo hoặc tồn lẻ; tồn đầu kỳ lẻ nhập được; tệp mẫu có cột
  Bán số lẻ.
- invariants.sql thêm I11: hàng không bật cờ thì tồn nguyên.

+++ packages/shared/src/schema/dashboard.ts
@@ -31,7 +31,7 @@ export const revenueChartItemSchema = z.object({
 export const topProductSchema = z.object({
   productId: z.string().uuid(),
   name: z.string(),
-  quantity: z.number().int(),
+  quantity: z.number(),
   revenue: z.number().int(),
   percentage: z.number(),
 })
@@ -39,8 +39,8 @@ export const topProductSchema = z.object({
 export const lowStockAlertSchema = z.object({
   productId: z.string().uuid(),
   name: z.string(),
-  currentStock: z.number().int(),
-  minStock: z.number().int(),
+  currentStock: z.number(),
+  minStock: z.number(),
   status: z.enum(['out', 'low']),
 })
